requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2

# Data Processing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1

# Database
# sqlite3==1.0.0
sqlalchemy==2.0.22

# LLM Integration
requests==2.31.0

# Embedding and Vector Search
sentence-transformers==2.2.2

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
tqdm==4.66.1

# Testing
pytest==7.4.2

# UI (for demo)
streamlit==1.27.2
plotly==5.17.0

# File handling
openpyxl==3.1.2